Removed unnecessary passing of isOwner prop
Because of permissions and access rules, if the user can see the edit button and navigate to the update page, they are the owner. No need to pass isOwner as query param on the way to the update page, and fine to just pass isOwner: true on the way from update to product.